deploy-requirements.txt
PyYAML
plumbum